Fix Sum op kernel bug for models with int8/uint8 input.
Fixes two bugs:
* The non-quantized sum kernel was being called for models with quantized
  inputs, resulting in incorrect results due to integer overflow.
* The optimized version of QuantizedMeanOrSum was not clamping values after
  executing the computation, resulting in overflow in some cases.

For the first issue, there was a problematic predicate for determining whether
to use quantization-aware kernels for the sum op. Before this change, inputs
had to be of type int8/uint8 and inputs/outputs had to have different scales
and zero points in order to qualify. The latter criteria has been removed.

One additional change worth noting (not quite sure it can be categorized as a
bug): before this change, only the reference version of QuantizedMeanOrSum was
being called. This change enables either the reference or optimized version to
be called depending on the specified kernel_type.

PiperOrigin-RevId: 538835574