Update list_physical_devices and list_logical_devices Argument notes
At present both the APIs list_physical_devices and list_logical_devices has an optional argument named device_type which takes a string as an argument. But When we pass a numeric value or strings other than expected('CPU' or 'GPU') both returns an empty list. This is not documented well.

Also the default value of the argument device_type is None but with default value both the APIs returns available physical/logical devices. This is not intuitive as for value None the APIs returning all the respective devices. This may create some confusion among users. Hence I am adding the note to the documentation explicitly to avoid any confusion.

The issue discussed in #57153 .