Fix tf.experimental.numpy.take_along_axis dtype bug and add a test for it.
tf.experimental.numpy.take_along_axis failed because the tf.where_v2 call in the function used the shape of the array which is always tf.int32. If indices passed as argument have a type that can't be safely coerced to int32, the function would fail. This PR fixes the issue and adds a test for it.

PiperOrigin-RevId: 539807296