Add basic debugging facilities for the TFLite Converter.
Adds the following functionalities:
1. MLIR before/after pass dump instrumentation.
2. Crash reproduction.
3. Pass timing.

There's a new `DebugOptions` message type for encapsulating debug related
parameters. In order to expose these options to the Python API, we include it
in the existing `TocoFlags` proto message.

The first two facilities are controlled by the new `mlir_dump_dir` option. If
specified, then both of the above features will be enabled and produce files
containing MLIR that is output to the specified directory. The former feature
outputs the MLIR before and after each pass. The latter will produce the
initial MLIR with additional pass metadata for additional debugging with tools
such as `tf-opt`.

For the before/after MLIR dump mechanism, the `mlir_dump_pass_re` option allows
the specification of a regex pattern to match against the pass names to filter
for specific passes if so desired.

Pass timing is controlled by the `mlir_enable_timing` option. Setting this
option to true will output statistics on pass latency, which can be useful for
profiling purposes.

PiperOrigin-RevId: 540103558