Use PrecisionConfig to determine whether TF32 is used on NVIDIA GPUs.
If the Precision for any input is HIGHEST or if the TF32 global variable is set to false, then TF32 will not be used. Otherwise TF32 will be used. In the future (hopefully soon), XLA will no longer use the TF32 global variable and only use the PrecisionConfig. This change is the first step to removing usage of the TF32 global variable.

Before, the PrecisionConfig was used for dots in some cases and never for convolutions. Now the PrecisionConfig is always used on NVIDIA GPUs, for both dots and convolutions.

Previously in cases where the PrecisionConfig was used, XLA was inconsistent whether HIGH was enough to disable TF32, or if HIGHEST was required. Which PrecisionConfig was required to disable TF32 depended on the codepath taken. Now HIGHEST is always required to disable TF32.

Previously, a ComputePrecision (a typedef-ed int) was passed to some StreamExecutor methods to indicate precision, with 0 being interpreted as TF32 and nonzero being interpreted as FP32. Such functions now take NumericOptions instead, which has an allow_tf32 field added. TensorFlow OpKernels are updated to pass the correct NumericOptions to StreamExecutor.

The "TensorFloat-32 will be used for the matrix multiplication" log message is now removed. It was not particularly useful and there was a TODO to remove it.

PiperOrigin-RevId: 540448127