Do not call cuCtxSetCurrent unnecessarily
The extra condition is not necessary and results in calling the `cuCtxSetCurrent` too often, which results in unnecessary host kernel scheduling overhead.

The thread local storage (tls) would contain a different ID, if some other code has run on the same thread and updated the context - the code comment (now removed) is misleading.

PiperOrigin-RevId: 540578636