[XLA:GPU] Fix inplace DUS emitter to handle reshapes correctly and reenable it.
This allows doing dynamic slice updates inplace for fusions in which all
outputs are potentially bitcasted dynamic slice updates that use the same
update shape, assuming that the parameters being updated have a unique
consumer. The parameters may also be bitcasted.

The previous attempt to do this did not compute update indices correctly, so
this also adds more numerical tests to ensure that all cases are well-covered.

PiperOrigin-RevId: 540592670