Fix build breakage resulting from name clash in generated Flatbuffer Java API.
The flatbuffer-generated Java API contained the following method:

 public static int createOperator(FlatBufferBuilder builder,
      long opcodeIndex,
      int inputsOffset,
      int outputsOffset,
      byte builtinOptionsType,
      int builtinOptionsOffset,
      int customOptionsOffset,    <--------------- This conflicts
      byte customOptionsFormat,
      int mutatingVariableInputsOffset,
      int intermediatesOffset,
      long customOptionsOffset,   <--------------- with this.
      long customOptionsSize) {

The Flatbuffer compiler is appending "Offset" to the field name to
get the Java parameter name for fields with vector types, such as the
"custom_options" in this schema, since it passes such fields by an
offset into the flatbuffer; but is not appending anything to the field
name for fields with scalar types, such as "custom_options_offset". This
then leads to a name clash between the parameters for those two fields.
This problem will occur in general whenever you have a non-scalar field
whose name, after appending "_offset", is the same as the name of a
scalar field. It is arguably a bug in the Flatbuffer compiler, but we
need to work around it.

The work-around here to just rename the field that caused the name clash.

PiperOrigin-RevId: 540907733