Support missing inputs and variable snapshots in PJRT execution related functions.
Also refactors a few functions to take in variable snapshots (i.e. the underlying Tensors that the variables point to) instead of VariableInfos.

This is needed for executing PjRtLoadedExecutables in the XlaRun op. Must-be-constants are stripped from the inputs to this op and this op also requires variable snapshots (i.e. variable values when compilation was first requested i.e. in XlaCompile) instead of the current variable values.

PiperOrigin-RevId: 540940440