Add a GTest matcher and printer for TFLite arrays.
The matcher will work with TfLite{Int,Float}Array values and pointers. The
pointer matching is added for convenience purposes as the majority of use cases
in TFLite are through pointers.

This also adds the necessary machinery to print TFLite arrays values when tests
fail. A side effect of that is that we can now use the `absl` string and format
utilities to print/debug the contents of arrays.

PiperOrigin-RevId: 541574408