Let CopyFusion pass support multi-output fusions.
Also allow an intermediate bitcast op between the fusion and the copy, and
allow it if there is only one copy.
Restrict running the pass to the entry computation. Inside while loop bodies,
fusing the copies can be wrong if they share the buffer with one of the fusion
parameters.

PiperOrigin-RevId: 541620589