[XLA:GPU] Check whether transpose fusions stay transpose fusions.
There are cases where due to forming multi-output fusions or merging fusions
with FusionMerger we turn fusions which are supported by the shared memory
transpose emitter into fusions that are not supported by the shared memory
transpose emitter anymore. This change tries to identify such cases, and in this
case we do not consider the fusion heroes to be compatible.

PiperOrigin-RevId: 542162676