Allow per-worker variables to be retrieved each time they are read, to ensure results are up-to-date.
Per-worker variables are different from other types of worker resources in that they are mutable. The caching used by default for other resources prevents up-to-date reads of per-worker variables, because the cached value will be returned when a new read is desired. This change introduces a subclass of RemoteValueImpl called RemoteVariable that does not use caching.

PiperOrigin-RevId: 542315482