[CMake] Add dependency for MhloTestAnalysis.
This patch declares LMHLOTransformsPassIncGen as a dependency to
MhloTestAnalysis. Without it, building on a single core would attempt to
compile `test_shape_component_analysis.cc` without having generated file
`transforms/passes.h.inc`.

This error is not encountered when building on multiple cores as the
distribution of the builds ensures that the file `transforms/passes.h.inc` is
generated before attempting to compile `test_shape_component_analysis.cc`.