Add a library that implements stable serialization of `xla::ifrt::XlaProgram`
This CL implements stable serialization and deserialization of `xla::ifrt::XlaProgram` using StableHLO. Currently, XLA programs passed to PjRt-IFRT contain C/MHLO, which provide no version compatibility and are not suitable for stable serialization. This CL leverages C/MHLO <-> StableHLO legalization to use StableHLO's portable serialization for `XlaProgram` serialization, while keeping the contract the same.

If the input module has ops that cannot be converted into StableHLO, the serialization will fail. This is expected since currently StableHLO is a smaller subset of MHLO. Also, it is intentional as users often make assumption that serialization is stable and build systems under such assumptions, so we might as well start from a smaller set that's serializable in a stable way and expand as necessary.

Since not all users of `xla::ifrt::XlaProgram` may want serialization, this is implemented as a standalone library that callers need to explicitly link in only if serialization is required. This is useful since the serialization/deserialization implementation requires several passes from StableHLO/CHLO/MHLO that are otherwise unnecessary for typical PjRt-IFRT users.

Deserialization requires someone to own the MLIR module and this is not possible today because `xla::ifrt::XlaProgram` contains a non-owning pointer (`mlir::ModuleOp`) only. Thus, `xla::ifrt::XlaProgram` was extended to optionally allow owning the `mlir::ModuleOp` (hidden to callers). This makes `xla::ifrt::XlaProgram` move-only, which is fine in my opinion.

In future, we may change PjRt-IFRT to accept StableHLO directly and let it handle StableHLO -> MHLO legalization. If and when that happens, this SerDes implementation can be simplified by serializing the StableHLO directly.

PiperOrigin-RevId: 543916457