Extract code for calculating the fusion launch dimensions into a separate class
The methods in HloFusionAnalysis are copied from the IrEmitterUnnested implementation, but have the references to LHLO stripped.
A future changelist will update the IrEmitterUnnested to reuse the methods in HloFusionAnalysis and remove the duplicate code.

The performance model uses HloFusionAnalysis to estimate the thread count for compute-bound fusions (e.g. reductions, transposes, loop fusions).
It falls back to the previous estimation if the fusion analysis fails (e.g. scatter)

PiperOrigin-RevId: 544028940