[XLA:GPU] Add --xla_gpu_redzone_padding_bytes flag.
Previously, the redzone allocator always added 16MiB per buffer (8MiB per
side).  A conv can have up to 6 buffers (input, filter, bias, side-input,
scratch, and output), and 96MiB is a lot of memory for smallish inference
workloads.  In particular, if the underlying BFCAllocator has a memory limit,
we can easily exceed it by adding 96MiB for a convolution.  This can cause
autotuning to fail.

This flag allows the user to specify a smaller redzone so they can avoid this
problem.

PiperOrigin-RevId: 545287511