NFC: Remove duplicate checks in GetProducerConsumerMultiOutputFusionCandidates.
The function currently checks every possible user of the producer with
`ShapesCompatibleForMultiOutputFusion`, rechecking the producer each time.
Instead, we can just check the producer once and exit early. Additionally, the
consumer checks are currently duplicated (`IsProducerConsumerMultiOutputFusible`
checks the candidate with `IsProducerMultiOutputFusible`, which
`GetProducerConsumerMultiOutputFusionCandidates` also does).

Ultimately, the only check that is left that depends on both the producer and
consumer is `ShapesCompatibleForMultiOutputFusion`, so we can just inline that.

Also, some drive-by cleanups in `gpu_fusible`.

PiperOrigin-RevId: 545402401