[XLA:GPU][PriorityFusion] Put back code size heuristic
We disabled this because the caching can't deal with fusions being merged. This
however still triggers exponential compile time and memory usage and it doesn't
look like it makes sense to address this with a cost model. Put it back for
testing, we may want to investigate how to bring the caching back later.

PiperOrigin-RevId: 545466777