Add additional debugging options to the TFLite Converter.
Adds four new debug-related options that enable printing MLIR to stdout:
* mlir_print_ir_before: Prints MLIR before each pass. Accepts a regex used for
  selectively matching against pass names.
* mlir_print_ir_after:  Prints MLIR after each pass. Accepts a regex used for
  selectively matching against pass names.
* mlir_print_ir_module_scope: Determines whether to print the top-level
  operation when printing IR.
* mlir_elide_elementsattrs_if_larger: Elides ElementsAttrs with number of
  elements larger than the provided integer.

The existing options are untouched in their functionality and control of
dumping MLIR to files is independent of that for printing to stdout. Both
can be done simultaneously if the user so desires.

A future CL will plumb the new options through the Python API.

PiperOrigin-RevId: 545875450