#tf-data-service RepeatedDatasetRecoversAndCompletes test verifies the elements.
Also update the repetition index for SnapshotSplitProvider. When
restored from the files, the repetition index is not set directly to
the repetition index of the latest split. It should be set to the
current repetition index, and returns `end_of_split` when the repetition
finishes. Otherwise, the iterator does not remember the right number of
repetitions and resets itself more than the expected times, resulting in
duplicate elements.

PiperOrigin-RevId: 546360505