Migrate ops fuzzers to FuzzTest format
Creates a set of fuzzers to FuzzTest style. One of the benefits is that
this will reduce storage size required for the fuzzers, because the
fuzzers will all be based off the same binary whereas the non-FuzzTest
fuzzers will have a binary per fuzzer. This storage requirement has
caused an issue on the OSS-Fuzz side:
https://github.com/google/oss-fuzz/issues/9792

Migrating to FuzzTest will make it easier to add new ops-specific
fuzzers since we won't have to worry about disk size. The existing
fuzzesr can be several GBs of storage and it quickly goes beyond what we
have available at OSS-Fuzz.

It would however be nice to re-use the existing fuzzer helper logic from
fuzz_session.h so I added a macro similar to what we have for libFuzzer
but simple for FuzzTest. I migrated a set of the existing fuzzers.

I figured it would be nice to do a soft transition, so we can remove the
existing fuzzers after the ones introduced in this PR has shown to run
for a while. I'm also curious in whether we will see a difference in
performance from the FuzzTest and libFuzzer harnesses.

Signed-off-by: David Korczynski <david@adalogics.com>