Add a proto version of `xla::ExecuteOptions`
`xla::CompileOptions` already has a proto version of it, but `xla::ExecuteOptions` does not. This CL adds the same `ToProto()`/`FromProto()` utility function for `xla::ExecuteOptions`.

It is worth noting that not all execute option fields are serializable (e.g., opaque `context`). For now, `ExecuteOptions::ToProto` returns an error if non-serializable options are set. In future, we can consider redesigning them to be serializable if needed.

PiperOrigin-RevId: 546897094