Fix index out of bounds access in cuda_dnn.cc
We might have removed one element from data_ptrs_vec, so accessing it with
sizeof...(Args)-1 can be out of bounds. What we actually want to do is to check
the last element.
Also apply several fixes for ClangTidy warnings.
Re-enable a test that exposed this bug when run in debug compilation mode.

PiperOrigin-RevId: 547096288