Add TfLiteInterpreterOptionsSetOpResolverExternalWithFallback.
This combines the effects of TfLiteInterpreterOptionsSetOpResolverExternal
and TfLiteInterpreterOptionsSetOpResolver, allowing use of an op resolver
that uses TfLiteRegistrationExternal for user-defined custom ops while still
continuing to use the builtin op resolver, which uses TfLiteRegistration objects
(that don't set the TfLiteRegistrationExternal field), for builtin ops.

This change required modifying the order in which we consider the fields of
CallbackOpResolver in c_api.cc so that we consider the _external fields first, before the other ones.  Previously the order didn't matter since these fields were mutually exclusive, but now that TfLiteInterpreterOptionsSetOpResolverExternal sets both the _external and regular fields, we want to consider the _external fields first, so that the op resolver callbacks that return TfLiteRegistrationExternal take precedence over the ones that return TfLiteRegistration, as specified in the documentation for the new API function TfLiteInterpreterOptionsSetOpResolverExternalWithFallback.

PiperOrigin-RevId: 547649220