Fix segfault in `XlaCallModule` shape inference
Entries to `xla_call_module_loaders_` were added eagerly with nullptr as values (to save a lookup on a miss), but this caused some entries to keep nullptr loaders if we fail to initialize the loader. This CL changes the logic so that we insert only non-nullptr loaders to the map.

Confirmed that the added MLIR test crashes without the fix.

PiperOrigin-RevId: 547656759