Revert: Attempt to be less restrictive in FusionCanShareBufferHint().
Initially I added a restriction that a fusion parameter should not be
(transitively) used by more than one fusion output. I added this restriction
because there was a test that was failing otherwise. After thinking about this
again, the real bug was just that I didn't check whether any of the users of the
fusion output for which we want to see whether it can share the buffer with the
fusion operand has some non-elementwise user.

PiperOrigin-RevId: 547702127