[xla:gpu] Improve stream assignment for concurrent regions
Before this change we assigned each kernel in a concurrent region to a different borrowed stream. We can instead assign one of the kernel to the capture stream in order to perform less synchronization.

For example, if we have two kernels in the concurrent region, we synchronize two borrowed streams:
   1
  / \
  2  3
  \ /
   1
After this change we only synchronize one stream:
   1
   |\
   1 2
   |/
   1

PiperOrigin-RevId: 547935901