Ensure there is a Model instance in standalone::Iterator's context
Currently, the model instance in the IteratorContext of a standalone::Iterator is null. To solve this, we need to create one manually when initializing the IteratorContext. Then, since RootDataset creates its own Model instance, we will replace it with the one passed from our IteratorContext. This is to make sure this instance is the one used down the chain and that we can also access it in our IteratorContext. By passing this model instance, we also ensure that when executing IteratorBase::InitializeBase we are creating a RootDataset node without a parent in the model.

PiperOrigin-RevId: 548201925