Modify the signature of `QuantizeModel` to accept the buffer directly instead of `ModelT`
Instead of accepting `ModelT` for quantization, it uses the raw model buffer as a parameter to `QuantizeModel` in order to support the case where the "offset buffer" exists.
Offset buffers are enabled by `_experimental_use_buffer_offset`.
This case happens usually when the model size is larger than 2GB for which the concept of offset buffer is introduced to avoid the flatbuffer's serialization limit of 2GB (see the previous commit that introduced this support for compiler import / export: https://github.com/tensorflow/tensorflow/commit/909b28382c31e3ada3872c11a767e64144a56db6)

PiperOrigin-RevId: 548599861