Add a ragged type Layout.
This amounts to adding a new type / one bit of information to Layout structure.

This should roughly translate to XLA SPMD's manual sharding notation; but details are not quite sorted out. There is some possibility to generalize RAGGED to partially ragged which is not supported by XLA SPMD, but for now (data-parallel as
the primary user), we don't overdesign the data structure for representing the partial ragged layout.

Python binding will be added in a followup CL.

- The typing of Layout class is generalized with enums.
- Parsing and serialization of string format is slightly updated.
- Protobuf message type is also updated.
- Fixed Status creation methods to the new absl syntax.

- Removed deps on LayoutProto in ConcatenateLayouts.

PiperOrigin-RevId: 548843187