Add GPU ID to FFT plan cache.
Previously, the CUDA FFT plans were cached based solely on FFT parameters.
However, this breaks in multi-GPU setups, since `cufftHandle`s are device-specific.
This led to garbage outputs when an FFT plan was attempted to be used on a
different device.

Added the GPU device ID to the plan cache.

Fixes #60926.

PiperOrigin-RevId: 549375830