[TF:PJRT] Some small fix ups to GPU CreateDevices behavior.
In unit tests, CreateDevices() are sometimes called repetitively. In this case, we reuse the existing PjRtClient already created and hosted on resource mgr singleton.  This is the same as legacy behavior since before PJRT, many GPU states are managed by a global singleton and we just use GetOrCreate to reuse the states if already created.

PiperOrigin-RevId: 549791661