Introduce version 8 of XlaCallModule.
Previously, XlaCallModule was running the shape refinement pass for all
compilations, even if the module did not use shape polymorphism.
Currently shape refinement changes the structure of the module,
through inlining and constant folding all integer operations.
This complicates debugging because the HLO dump is very different
than the one from JAX native executions.

Starting with version 8, we run shape refinement only
if the module contains a boolean module attribute
jax.uses_shape_polymorphism=true. I think it makes sense
to put this flag as a module attribute, rather than
as a TF op attribute, because the same processing will
be needed when the module is executed from JAX.

This attribute is not yet populated by the JAX exporter.

As part of this change we moved the error check for the
number of invocation arguments from RefineDynamicShapes
to LoadAndPreprocessModule. This required adding a couple
more arguments to the loader constructor.

PiperOrigin-RevId: 549973693