roll forward with fix: bincount op for sparse tensors
New:

Remove code that counted missing zeros in sparse tensors.

Document that missing zeros in the sparse tensor input are not counted. This supports cases such as `0` in the values tensor indicates that index/id `0` is present and an implicit zero indicates that no index/id is present. Keras `CategoryEncoding` is an example of this case.

Previous:

Fix a maxlength related issue in the C++ kernel.

In the Python for bincount for spare inputs, remove an obsolete codeblock and handle the case when the input is only missing zeros.

Improve test coverage.

PiperOrigin-RevId: 550074654