[XLA:GPU] Fix output transpose fusion in Triton GEMM.
To set up the dot output fusion LHS parameter instructions need to be searched in the generated dot input fusion, not the original computation. Dimension orders need to be stored for all inputs of the fused instructions, not just the fused ones.
A test was updated to demonstrate these problems.

PiperOrigin-RevId: 550509732