Fix a bug in the RemoveReshapeBeforeFullyConnected pattern.
Fixes a bug in how RemoveReshapeBeforeFullyConnected pattern is implemented in the converter. The bug is that this pattern is missing a check-

Consider-
```cpp
func.func @RemoveReshapeBeforeFullyConnectedExpandDims0(%arg0: tensor<128x64xf32>, %arg1: tensor<32x64xf32>, %arg2: tensor<32xf32>) -> tensor<128x32xf32> {
  %cst = arith.constant dense<[1, 128, 64]> : tensor<3xi32>
  %0 = "tfl.reshape"(%arg0, %cst) : (tensor<128x64xf32>, tensor<3xi32>) -> tensor<1x128x64xf32>
  %1 = "tfl.fully_connected"(%0, %arg1, %arg2) {fused_activation_function = "NONE", keep_num_dims = false, weights_format = "DEFAULT"} : (tensor<1x128x64xf32>, tensor<32x64xf32>, tensor<32xf32>) -> tensor<128x32xf32>
  func.return %1 : tensor<128x32xf32>
}
```
ends-up becoming the following, after applying the pattern-
```cpp
   %1 = "tfl.fully_connected"(%arg0, %arg1, %arg2) {fused_activation_function = "NONE", keep_num_dims = false, weights_format = "DEFAULT"} : (tensor<128x64xf32>, tensor<32x64xf32>, tensor<32xf32>) -> tensor<128x32xf32>
   return %1 : tensor<128x32xf32>
```

Since `keep_num_dims` is false, the FC op output dimensions will always be of rank-2. Thus any reshape to increase the rank of the input(weight) will not effect the FC output shape. Thus, it is safe to eliminate such reshapes prior to FC.

So, the pattern must be checking if the reshape is actually increasing the rank of the input, not decrease.

PiperOrigin-RevId: 550574670