Rename "nvidia" configs to "cuda" configs, adjust utility scripts, fix typos
Utility scripts and bats tests no longer reference TFCI variables, since they
won't be available inside of the docker environment. I think this makes sense
and decouples the scripts from their environment -- now they merely expect to be
run from the root of the tensorflow directory.

I also renamed "nvidia" to "cuda" to fit with some other changes we're doing,
and fixed a typo in the env files where "--gpu" should not have been there.

I also found out while working on this that "enviroment variables" MUST be
key/value _string_ pairs and cannot be arrays. All the arrays I'm using in
the env files work as intended, but do not appear in the output from the "env"
bash command. This means we cannot e.g. pass those array values into a docker
container. I don't think this is a problem, but it's worth noting and I'll
make mention of it when I start documenting these further.

PiperOrigin-RevId: 550609388