Potential improvement opportunity to eliminate extra transpose op.
This CL will add patterns to fold Transpose and FC to covert into a BMM, like below-

FC(lhs, Transpose(rhs)) -> BMM(lha, rhs, false, false)

The right thing to do in this pattern will be to apply the pattern only if keep_num_dims==True. Because, if the output rank is less-than the input rank, it means `keep_num_dims` has reduced the output. But checking for rank will improve the coverage. This pattern will now work

PiperOrigin-RevId: 551297769