[FTLite] Fix pthreadpool CMake integration
With the changes introduced in 9c3e858 it is no longer possible to use a
prebuilt version of pthreadpool. The only options are download it or
specify a folder where it was downloaded.

Add SYSTEM_PTHREADPOOL as a new option that triggers find_library() and
fails if the library can't be found.