Register Iterations automatically when reporting a time
In MultipleIterationsAutoScaler, an Iteration had to be registered before starting to report processing or target processing times. However, there can be cases where the dispatcher is restarted but the Iteration is not registered again, causing the times' reporting to fail. To solve this, we allow reporting times even if the specified Iteration has not been registered before. If it does not exist already, we create one.

PiperOrigin-RevId: 551631559