Limit the increase of number of tf.data service workers
The received target processing times can be very low if there is a bottleneck in the input pipeline, which can cause the estimated optimal number of workers to be unreasonably high. This is because the recorded gap times in model.cc are very low. To solve this, we limit the increase of the number of workers with respect to the current number of registered workers in the tf.data service cluster. This way, we increase the number of workers gradually until gap times start to be higher.

At first, we allow increases of at most 4x the current number of workers, and then we limit the increase to at most 500 workers.

PiperOrigin-RevId: 552822340