Return average gap time for experimental target time computation
Very low values are currently being returned for the target time estimate. As an attempt to obtain more reasonable values for input-bottleneck detection, we return the average of gap times.

Another issue is that subtracting target time sigmas can cause the estimated target time to be negative. To solve this, we limit the target time estimate to be at least 0.

PiperOrigin-RevId: 552874042