Fix output variable updates (shape and type) when PJRT is enabled.
The XlaClient code has always used the ResourceUpdate's type and shape from the ComputationResult [1]. In the case of PJRT, the resource variable updates use the Var's type and shape. The Var's type and shape should be the same as the ResourceUpdate's type and shape, but because the Var is uninitialized for the first time, the type and shape don't make sense (eg. type is DT_INVALID).

PiperOrigin-RevId: 552973251