[IFRT] More plumbing of MemoryKind between JAX/PjRt and IFRT
This change introduces more `MemoryKind` plumbing between JAX and IFRT (in the
JAX backend) and IFRT and PjRt (in the PjRt-IFRT implementation).

The `MemoryKind` is plumbed from an executable to the JAX sharding of output
JAX arrays only when the environment variable
`JAX_FETCH_MEMORY_KIND_ON_EXECUTABLE` has a non-empty string. Since this is
undefined by default, this CL is not expected to set the `memory_kind` of JAX
shardings some value other than `None`.

PiperOrigin-RevId: 553354312