Prepare to remove obsolete codeblock in bincount for dense tensors.
Prepare to use newer DenseBincount kernel instead of older Bincount kernel for
tf.math.bincount on dense tensor inputs, but do not enable this until after
a forward compatibility window.
The shape function for the DenseBincount kernel now returns "vector with unknown size" or "matrix with unknown size" instead of "unknown" when the shape is not known exactly.
Added tests.
Documentation clarifications.

PiperOrigin-RevId: 553595278