Remove capture_test_logs helper
This is not useful at the moment. It adds about 35 minutes to our builds (wow!),
adds a bunch of junk log lines to the end of the build, and the uploaded logs
don't include FAILED_TO_BUILD records, which would just add to the confusion.
If it's deemed useful again later, we can put it back.

PiperOrigin-RevId: 554572172