Fix undefined CUDA_VERSION in gemm_rewrite.
Preprocessor macro CUDA_VERSION was used in gemm_rewriter.cc and gemm_rewrite_test.cc, but since cuda.h was not included, all #if statements with this macro returned false. This caused FP8 tests to be skipped and caused FP8 to be used on versions of CUDA that did not support it.

PiperOrigin-RevId: 554704361