[Linaro:ARM_CI] Run build tests
The build tests and similar have an empty language tag and so will
not run when test_lang_filters has a positive language choice, so
change it to use a negative choice to exclude the python based tests
that are tested in a different script.
Also remove use of python venv that is no longer useful.