Add additional TF-NumPy methods to WeakTensor.
In TF-NumPy, ndarray is introduced as an alias of tf.Tensor. A few additional methods that don't exist in tf.Tensor are added to ndarray class in np_array_ops: http://shortn/_FNWKMUHNan. Look for 'setattr' in np_array_ops to see all the methods that are added to ndarray class.

PiperOrigin-RevId: 556078735