When converting control to data, only consider edges that existed beforehand.
Before the control-to-data pass, all ops that need to be waited on before the
next iteration have a control dependency that goes to the "fetch" terminator.
The control to data pass is supposed to convert all these dependencies to
data edges (and make them into separate args/results to the while loop body,
so we get parallelism). However, what it did so far is to throw away all
control dependencies to the "fetch", and then recreate them, only considering
which resources are touched by which ops, across iterations. This caused
extraneous data dependencies to be created, even for ops that don't need them.
(For example, some "Delete" ops only use a resource to retrieve a constant
 key, so nothing ever needs to depend on the "Delete". Side effect analysis
 is aware of this, but control-to-data wasn't.)

This change now changes this behavior so we only add a new (data) dependency
if the op in question had a (transitive) control dependency (to the fetch) in
the first place.

PiperOrigin-RevId: 556841791