Convert "build/" dir to "build_output" dir and make it configurable
I originally used "build/" as the canonical output directory for all build
scripts, but that collided with a "BUILD" file in the TF root directory on
Windows. That was too confusing to handle, so I added a TFCI_OUTPUT_DIR
variable and set it to default to build_output instead, both of which are
additionally easier to search for. Just "build" is overloaded.

PiperOrigin-RevId: 556858432