Fix logic in grouped convolution slicing.
The slicing for the filter and bias was done incorrectly, using filter[IC] as the slice size, when in reality it should have been OC/(input[IC]/filter[IC]).
Made the implementation a bit more verbose for clarity since we are slicing 4 different elements along different dimensions with different slice sizes.

PiperOrigin-RevId: 557255842