Add `use_default_shell_env = True` to `ctx.actions.run`
This is required e.g. for `_generate_op_reg_offsets_impl`
which otherwise will ommit e.g. `LD_LIBRARY_PATH` which then fails to
find a (system-)libprotobuf.so when generating
`//tensorflow/python:math_ops_reg_offsets`

This is in line with other usages of this function to generate files
using external tools.