[pjrt] Add a kernel for 4x4 int16 transpose
Similar to the 8x8 int4 transpose, this is annoying because each load only occupies half the SSE vectors.

However, we can quickly zip together pairs to make a 2x8 matrix.
A small number of SIMD-wise zips results in the elements appearing in transposed order.

Finally, we can store the elements using `_mm_storeu_si64` with a potential shuffle for odd memory locations.

PiperOrigin-RevId: 557993476