[PJRT C API] Split profiler related C APIs to its own struct, and only initialize it in tpu_tracer.cc.
The goal is to move non-PJRT C API initializations from PJRT plugin common path to TPU specific code path.

Add a tsl_status_helper because status_helper depends on ExecutorApiFn. Also move the dependency of proto_helper to where it is needed to avoid MACRO definition conflict.

The only behavior change with this change is when initializing profiler C APIs fails. Before this change, a tpu client will not be created. After this change, it will only fail when the user uses profiler.

PiperOrigin-RevId: 558271842