Remove deprecated "distributions" package from exports.
It has been deprecated since 2019, and is now triggering warnings on
TensorFlow startup.  It does, however, still need to be included in
`python/__init__.py` so that it can be found by `tf.compat.v1.distributions`,
which several projects depend on.

We will remove the entire package from TensorFlow as a follow-up.  But
at the moment, doing so would break too many other projects.

PiperOrigin-RevId: 558816905