Support quantized i64 during flatbuffer import
Sometimes tflite flatbuffer represent `i32` quantized values as `i64`s. In these cases we should truncate down to a lower bit width to avoid creating illegal types.

We check the bitwidth of the type at load time and pick a power-of-2 bit
width where the value can be safely truncated.

PiperOrigin-RevId: 558958449