Move kernel gen jit outlining to the whole function level.
Before, the outlining would match a compatible operation and produce a
corresponding jit region. However, we end up always outlining the whole bode of
the function anyway. With this change, we unconditionally do this independent of
the contained operations. This enables the use of kernel_gen with HLO instead
of just TF dialects.

PiperOrigin-RevId: 559068652