Add QuantizeWeightsPass
This pass covers const weights used by TF/TF XLA ops with f32/bf16
precision. After quantization, a constant will be replaced by a
quantized const + ops for dequantize combination. In the quantize op,
without an identity op, const will be folded. In the dequantize op, a
cast op is required to buffer between the new data type and the original
type. If the next op supports the new data type, then the type will be
propagated which will be done by the follow-up CLs.

PiperOrigin-RevId: 559169311