PR #5103: Force the cublasCreate call before we start the graph.
Imported from GitHub PR https://github.com/openxla/xla/pull/5103

Calling cublasCreate() need locks and can sync the context. If we have kernels that are already running and are waiting for something, like a NCCL kernel, we end up with deadlock. [Section 20.5.1 (Concurent Execution) of the CUDA programing guide
](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#concurrent-execution) talk about that.

This small PR is only a partial fix for the old runtime.
There is 2 cases not covered:
- the new runtime
- a rare case, imagine 2 xla graph. The first one is started but don't need cublas. It fill the GPU with long kernel and return async. Then you start another graph. Possibly compiling it. Then executing it. The first graph is still running here!
Now, we still call cublasCreate() while the other graph is still running. So the old runtime still risk a deadlock.
@ezhulenev I think this interest you.

We need to also check if cudnn need such fix.
Copybara import of the project:

--
6f864b6cc1685a33c2f1c7cd7fec0d5461847f00 by Frederic Bastien <fbastien@nvidia.com>:

Force the cublasCreate call before we start the graph. Not doing it can mcause the init to happen when there are kernels already running and this break the cuda programing contract and cause hangs in some cases.

Merging this change closes #5103

PiperOrigin-RevId: 559887430