Fix delegation of VarHandle nodes for unsupported datatype
VarHandle nodes don't have a type associated with it. We rely on a ReadVariable/AssignVariable to figure out the data type. When visiting VarHandle, we currently incorrectly assume we can always handle it, but it really depends on the data type, which we only know later.

So, instead of assuming we can handle it, we assume we cannot handle it. And after the first pass of VisitNode, the ReadVariable/AssignVariable has informed us of the datatype, and we know if we can handle the variable. We check all VarHandle nodes and only update nodes_to_delegate based on whether we can handle the data type.

PiperOrigin-RevId: 559932750