Allow i8->i32 convert op for zero point constants in `ComposeUniformQuantizedDotGeneralOpWithTwoQuantizedActivations`.
Currently the `mlir::odml::ComposeUniformQuantizedDotGeneralOpWithTwoQuantizedActivations` assumes the zero point values are constant folded into i32 values.
However, there are use cases where the zero points don't get folded and there is a separate `stablehlo.convert` op to cast it to i32.
This change allows a variation of the pattern to allow the case above.

PiperOrigin-RevId: 559933356