Add tests and cleanup for multi-platform support in XlaCallModule.
XlaCallModule already had support for multi-platform modules, even
though JAX serialization does not yet export such modules. Here we
add a test for combining multi-platform with shape polymorphism.

We also clean up the implementation. Previously the platform index
was handled in AddMainWrapper, along with the dimension variables.
That function should go away in 1 month when we have 6 months since we stopped using dimension variables in the API for XlaCallModule.

We implemented a simpler way to set the platform index in a new function SetPlatformIndex.

This is preparation for launching multi-platform support in JAX serialization.

PiperOrigin-RevId: 560934572