Align reshape and squeeze integer array types with flatbuffer data type
The integer-based arrays in the TFLite flatbuffer schema for ReshapeOptions and SqueezeOptions result in a generated `flatbuffers::Vector<int32_t>` for C++. Their BuiltInOpData counterparts of `TfLiteReshapeParams` and `TfLiteSqueezeParams` have `int[]`. On some of downstream TFLM's supported targets, `int32_t` is a `long int` instead of an `int`, even though they are both 4 bytes. This produces a compiler error for these platforms due to the mismatched types.

This change aligns the integer arrays in the `TfLiteReshapeParams` and `TfLiteSqueezeParams` to also be `int32_t` based.

PiperOrigin-RevId: 561754961