Add configurable retries for SetServerDef, for fault tolerance amidst preemptions.
There's a short period during ParameterServerStrategy initialization / cluster connection in which worker preemptions will lead to UnavailableErrors from CreateContext calls. This adds configurable retries to SetServerDef so that a single connection failure does not stop the whole job. Retries will be enabled as the default behavior for PSS in a followup change.

PiperOrigin-RevId: 565214961