[xla] Extend collective-pipeliner to support the pipelining of Send and Recv chains.
Extend collective-pipeliner to support the backward pipelining of Recv-done,
which has the effect of pipelining a Send and Recv chain.

Extend the HLO verifier to support multiple Send and Recv pairs using the same
channel IDs, which is a new pattern created by pipelining Send and Recv chains.

Add a test.

PiperOrigin-RevId: 565400560